refactor : QRModal -> ./QR/QRModal로 이동, Divider 별도로 생성
