+++ src/components/Main/Common/Divider.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import styled from 'styled-components';
+import { Divider } from 'antd';
+
+export const StyledDivider = styled(Divider)`
+    border-color: black;
+`;
+
+export const DividerWrapper = styled.div`
+    width: 100%;
+    margin: 0;
+`;
