+++ src/pages/MyPage/mypage.js
@@ -1,48 +1,16 @@
 import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import axios from 'axios';
 import styled from 'styled-components';
-import { Divider } from 'antd';
 import Header from '../../components/Main/Common/Header';
 import SectionTitle from '../../components/MyPage/SectionTitle';
 import MyPageSection from '../../components/MyPage/MyPageSection';
 import UserInfo from '../../components/MyPage/MyUser/UserInfo';
-import ReservationInfo from '../../components/MyPage/MyReservation/ReservationInfo';
+import { ReservationInfo, NoReservationInfo } from '../../components/MyPage/MyReservation/ReservationInfo';
 import ReservationHead from '../../components/MyPage/MyReservation/ReservationHead';
 import ReservationBody from '../../components/MyPage/MyReservation/ReservationBody';
-import QRModal from '../../components/MyPage/QRModal';
+import QRModal from '../../components/MyPage/QR/QRModal';
 import { MyReservationAPI } from '../../api/MyPage/MyPageAPI';
-
-import { getCookie } from '../../utils/cookie';
-
-const NoReservation = styled.div`
-    position: relative;
-
-    margin: 50px auto; /*페이지 중앙에 나타나토록 설정*/
-`;
-
-const StyledDivider = styled(Divider)`
-    border-color: black;
-`;
-
-const DividerWrapper = styled.div`
-    width: 100%;
-    margin: 0;
-`;
-
-const SectionTitleWrapper = styled.div`
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-`;
-
-const MoreButton = styled.div`
-    cursor: pointer;
-    color: black;
-    font-family: 'Kanit';
-    font-size: 15px;
-    padding-right: 10px;
-`;
+import { StyledDivider, DividerWrapper } from '../../components/Main/Common/Divider';
 
 const MyPage = () => {
     const navigate = useNavigate();
@@ -127,7 +95,7 @@ const MyPage = () => {
                                 </ReservationInfo>
                             </QRModal>
                         ) : (
-                            <NoReservation>진행 중인 예약 내역이 없습니다.</NoReservation>
+                            <NoReservationInfo />
                         )
                     ) : (
                         <div></div>
@@ -139,3 +107,17 @@ const MyPage = () => {
 };
 
 export default MyPage;
+
+const SectionTitleWrapper = styled.div`
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+`;
+
+const MoreButton = styled.div`
+    cursor: pointer;
+    color: black;
+    font-family: 'Kanit';
+    font-size: 15px;
+    padding-right: 10px;
+`;
